moved fixtures and functions into conftest
